refactor: Extract hook for behavior to pop navigation and notify if data is missing
